new source 27032017

+++ lib/include/Mail/Queue/Container/mdb2.php
@@ -1,22 +1,43 @@
 <?php
 /* vim: set expandtab tabstop=4 shiftwidth=4: */
-// +----------------------------------------------------------------------+
-// | PEAR :: Mail :: Queue :: MDB2 Container                              |
-// +----------------------------------------------------------------------+
-// | Copyright (c) 1997-2004 The PHP Group                                |
-// +----------------------------------------------------------------------+
-// | This source file is subject to version 3.0 of the PHP license,       |
-// | that is bundled with this package in the file LICENSE, and is        |
-// | available at through the world-wide-web at                           |
-// | http://www.php.net/license/3_0.txt.                                  |
-// | If you did not receive a copy of the PHP license and are unable to   |
-// | obtain it through the world-wide-web, please send a note to          |
-// | license@php.net so we can mail you a copy immediately.               |
-// +----------------------------------------------------------------------+
-// | Author: Lorenzo Alberton <l.alberton at quipo.it>                    |
-// +----------------------------------------------------------------------+
-//
-// $Id: mdb2.php,v 1.9 2007/01/20 10:51:25 quipo Exp $
+/**
+ * +----------------------------------------------------------------------+
+ * | PEAR :: Mail :: Queue :: MDB2 Container                              |
+ * +----------------------------------------------------------------------+
+ * | Copyright (c) 1997-2008 Lorenzo Alberton                             |
+ * +----------------------------------------------------------------------+
+ * | All rights reserved.                                                 |
+ * |                                                                      |
+ * | Redistribution and use in source and binary forms, with or without   |
+ * | modification, are permitted provided that the following conditions   |
+ * | are met:                                                             |
+ * |                                                                      |
+ * | * Redistributions of source code must retain the above copyright     |
+ * |   notice, this list of conditions and the following disclaimer.      |
+ * | * Redistributions in binary form must reproduce the above copyright  |
+ * |   notice, this list of conditions and the following disclaimer in    |
+ * |   the documentation and/or other materials provided with the         |
+ * |   distribution.                                                      |
+ * | * The names of its contributors may be used to endorse or promote    |
+ * |   products derived from this software without specific prior written |
+ * |   permission.                                                        |
+ * |                                                                      |
+ * | THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS  |
+ * | "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT    |
+ * | LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS    |
+ * | FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE       |
+ * | COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,  |
+ * | INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, |
+ * | BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;     |
+ * | LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER     |
+ * | CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT   |
+ * | LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN    |
+ * | ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE      |
+ * | POSSIBILITY OF SUCH DAMAGE.                                          |
+ * +----------------------------------------------------------------------+
+ * | Author: Lorenzo Alberton <l.alberton at quipo.it>                    |
+ * +----------------------------------------------------------------------+
+ */
 
 /**
  * Storage driver for fetching mail queue data from a PEAR::MDB2 database
@@ -24,15 +45,27 @@
  * This storage driver can use all databases which are supported
  * by the PEAR MDB2 abstraction layer.
  *
- * @author   Lorenzo Alberton <l.alberton at quipo.it>
+ * PHP Version 4 and 5
+ *
+ * @category Mail
  * @package  Mail_Queue
- * @version  $Id: mdb2.php,v 1.9 2007/01/20 10:51:25 quipo Exp $
+ * @author   Lorenzo Alberton <l dot alberton at quipo dot it>
+ * @version  CVS: $Id: mdb2.php 303870 2010-09-29 16:25:34Z till $
+ * @license  http://www.opensource.org/licenses/bsd-license.php The BSD License
+ * @link     http://pear.php.net/package/Mail_Queue
  */
 require_once 'MDB2.php';
 require_once 'Mail/Queue/Container.php';
 
 /**
  * Mail_Queue_Container_mdb2
+ * 
+ * @category Mail
+ * @package  Mail_Queue
+ * @author   Lorenzo Alberton <l dot alberton at quipo dot it>
+ * @version  Release: @package_version@
+ * @license  http://www.opensource.org/licenses/bsd-license.php The BSD License
+ * @link     http://pear.php.net/package/Mail_Queue
  */
 class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
 {
@@ -42,7 +75,9 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      * Reference to the current database connection.
      * @var object PEAR::MDB2 instance
      */
-    var $db;
+    var $db = null;
+
+    var $errorMsg = 'MDB2::query() failed: "%s", %s';
 
     /**
      * Table for sql database
@@ -56,12 +91,20 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
     var $sequence = null;
 
     // }}}
+    // {{{ __construct()
+
+    function __construct($options)
+    {
+        return $this->Mail_Queue_Container_mdb2($options);
+    }
+
+    // }}}
     // {{{ Mail_Queue_Container_mdb2()
 
     /**
-     * Contructor
+     * Constructor
      *
-     * Mail_Queue_Container_mMDB2:: Mail_Queue_Container_mdb2()
+     * Mail_Queue_Container_mdb2()
      *
      * @param mixed $options    An associative array of connection option.
      *
@@ -82,15 +125,88 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
         if (!empty($options['pearErrorMode'])) {
             $this->pearErrorMode = $options['pearErrorMode'];
         }
-        $this->db =& MDB2::connect($options['dsn'], true);
-        if (PEAR::isError($this->db)) {
-            return new Mail_Queue_Error(MAILQUEUE_ERROR_CANNOT_CONNECT,
-                $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
-                'MDB2::connect failed: '. MDB2::errorMessage($this->db));
-        } else {
-            $this->db->setFetchMode(MDB2_FETCHMODE_ASSOC);
+        $dsn = array_key_exists('dsn', $options) ? $options['dsn'] : $options;
+        $res = $this->_connect($dsn);
+        if (PEAR::isError($res)) {
+            return $res;
         }
         $this->setOption();
+    }
+
+    // }}}
+    // {{{ _connect()
+
+    /**
+     * Connect to database by using the given DSN string
+     *
+     * @param mixed &$db DSN string | array | MDB2 object
+     *
+     * @return boolean|PEAR_Error on error
+     * @access private
+     */
+    function _connect(&$db)
+    {
+        if (is_object($db) && is_a($db, 'MDB2_Driver_Common')) {
+            $this->db = &$db;
+        } elseif (is_string($db) || is_array($db)) {
+            include_once 'MDB2.php';
+            $this->db =& MDB2::connect($db);
+        } elseif (is_object($db) && MDB2::isError($db)) {
+            return new Mail_Queue_Error(MAILQUEUE_ERROR_CANNOT_CONNECT,
+                $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
+                'MDB2::connect failed: '. $this->_getErrorMessage($this->db));
+        } else {
+            return new Mail_Queue_Error(MAILQUEUE_ERROR_CANNOT_CONNECT,
+                $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
+                'The given dsn was not valid in file '. __FILE__ . ' at line ' . __LINE__);
+        }
+        if (PEAR::isError($this->db)) {
+            return $this->db;
+        }
+        return true;
+    }
+
+    // }}}
+    // {{{ _checkConnection()
+
+    /**
+     * Check if there's a valid db connection
+     *
+     * @return boolean|PEAR_Error on error
+     */
+    function _checkConnection() {
+        if (!is_object($this->db) || !is_a($this->db, 'MDB2_Driver_Common')) {
+            $msg = 'MDB2::connect failed';
+            if (PEAR::isError($this->db)) {
+                $msg .= $this->_getErrorMessage($this->db);
+            }
+            return new Mail_Queue_Error(MAILQUEUE_ERROR_CANNOT_CONNECT,
+                $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__, $msg);
+        }
+        return true;
+    }
+
+    /**
+     * Create a more useful error message from DB related errors.
+     *
+     * @access private
+     *
+     * @param PEAR_Error $errorObj A PEAR_Error object.
+     *
+     * @return string
+     */
+    function _getErrorMessage($errorObj)
+    {
+        if (!Pear::isError($errorObj)) {
+            return '';
+        }
+        $msg   = ': ' . $errorObj->getMessage();
+        $debug = $errorObj->getDebugInfo();
+
+        if (!empty($debug)) {
+            $msg .= ", DEBUG: {$debug}";
+        }
+        return $msg;
     }
 
     // }}}
@@ -104,6 +220,10 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      */
     function _preload()
     {
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
+        }
         $query = 'SELECT * FROM ' . $this->mail_table
                 .' WHERE sent_time IS NULL AND try_sent < '. $this->try
                 .' AND time_to_send <= '.$this->db->quote(date('Y-m-d H:i:s'), 'timestamp')
@@ -113,7 +233,7 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
         if (PEAR::isError($res)) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_QUERY_FAILED,
                 $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
-                'MDB2::query failed - "'.$query.'" - '.MDB2::errorMessage($res));
+                sprintf($this->errorMsg, $query, $this->_getErrorMessage($res)));
         }
 
         $this->_last_item = 0;
@@ -123,8 +243,11 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
             if (!is_array($row)) {
                 return new Mail_Queue_Error(MAILQUEUE_ERROR_QUERY_FAILED,
                     $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
-                    'MDB2::query failed - "'.$query.'" - '.MDB2::errorMessage($res));
+                    sprintf($this->errorMsg, $query, $this->_getErrorMessage($res)));
             }
+
+            $delete_after_send = (bool) $row['delete_after_send'];
+
             $this->queue_data[$this->_last_item] = new Mail_Queue_Body(
                 $row['id'],
                 $row['create_time'],
@@ -136,7 +259,7 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
                 $this->_isSerialized($row['recipient']) ? unserialize($row['recipient']) : $row['recipient'],
                 unserialize($row['headers']),
                 unserialize($row['body']),
-                $row['delete_after_send'],
+                $delete_after_send,
                 $row['try_sent']
             );
             $this->_last_item++;
@@ -169,13 +292,9 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
     function put($time_to_send, $id_user, $ip, $sender,
                 $recipient, $headers, $body, $delete_after_send=true)
     {
-        if (!is_object($this->db) || !is_a($this->db, 'MDB2_Driver_Common')) {
-            $msg = 'MDB2::connect failed';
-            if (PEAR::isError($this->db)) {
-                $msg .= ': '.$this->db->getMessage();
-            }
-            return new Mail_Queue_Error(MAILQUEUE_ERROR_CANNOT_CONNECT,
-                $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__, $msg);
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
         }
         $id = $this->db->nextID($this->sequence);
         if (empty($id) || PEAR::isError($id)) {
@@ -201,7 +320,7 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
         if (PEAR::isError($res)) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_QUERY_FAILED,
                 $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
-                'MDB2::query failed - "'.$query.'" - '.MDB2::errorMessage($res));
+                sprintf($this->errorMsg, $query, $this->_getErrorMessage($res)));
         }
         return $id;
     }
@@ -218,6 +337,10 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      */
     function countSend($mail)
     {
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
+        }
         if (!is_object($mail) || !is_a($mail, 'mail_queue_body')) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_UNEXPECTED, __FILE__, __LINE__);
         }
@@ -229,7 +352,7 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
         if (PEAR::isError($res)) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_QUERY_FAILED,
                 $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
-                'MDB2::query failed - "'.$query.'" - '.MDB2::errorMessage($res));
+                sprintf($this->errorMsg, $query, $this->_getErrorMessage($res)));
         }
         return $count;
     }
@@ -246,6 +369,10 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      */
     function setAsSent($mail)
     {
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
+        }
         if (!is_object($mail) || !is_a($mail, 'mail_queue_body')) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_UNEXPECTED,
                 $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
@@ -259,7 +386,7 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
         if (PEAR::isError($res)) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_QUERY_FAILED,
                 $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
-                'MDB2::query failed - "'.$query.'" - '.MDB2::errorMessage($res));
+                sprintf($this->errorMsg, $query, $this->_getErrorMessage($res)));
         }
         return true;
     }
@@ -276,14 +403,26 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      */
     function getMailById($id)
     {
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
+        }
         $query = 'SELECT * FROM ' . $this->mail_table
                 .' WHERE id = '   . (int)$id;
         $row = $this->db->queryRow($query, null, MDB2_FETCHMODE_ASSOC);
-        if (PEAR::isError($row) || !is_array($row)) {
+        if (PEAR::isError($row)) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_QUERY_FAILED,
                 $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
-                'MDB2: query failed - "'.$query.'" - '.$row->getMessage());
+                sprintf($this->errorMsg, $query, $this->_getErrorMessage($row)));
         }
+        if (!is_array($row)) {
+            return new Mail_Queue_Error(MAILQUEUE_ERROR_QUERY_FAILED,
+                $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
+                sprintf($this->errorMsg, $query, 'no such message'));
+        }
+
+        $delete_after_send = (bool) $row['delete_after_send'];
+
         return new Mail_Queue_Body(
             $row['id'],
             $row['create_time'],
@@ -295,9 +434,30 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
             $this->_isSerialized($row['recipient']) ? unserialize($row['recipient']) : $row['recipient'],
             unserialize($row['headers']),
             unserialize($row['body']),
-            $row['delete_after_send'],
+            $delete_after_send,
             $row['try_sent']
         );
+    }
+
+    /**
+     * Return the number of emails currently in the queue.
+     *
+     * @return mixed An int, or Mail_Queue_Error on failure.
+     */
+    function getQueueCount()
+    {
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
+        }
+        $query = 'SELECT count(*) FROM ' . $this->mail_table;
+        $count = $this->db->queryOne($query);
+        if (PEAR::isError($count)) {
+            return new Mail_Queue_Error(MAILQUEUE_ERROR_QUERY_FAILED,
+                $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
+                sprintf($this->errorMsg, $query, $this->_getErrorMessage($count)));
+        }
+        return (int) $count;
     }
 
     // }}}
@@ -313,6 +473,10 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
      */
     function deleteMail($id)
     {
+        $res = $this->_checkConnection();
+        if (PEAR::isError($res)) {
+            return $res;
+        }
         $query = 'DELETE FROM ' . $this->mail_table
                 .' WHERE id = ' . $this->db->quote($id, 'text');
         $res = $this->db->query($query);
@@ -320,7 +484,7 @@ class Mail_Queue_Container_mdb2 extends Mail_Queue_Container
         if (PEAR::isError($res)) {
             return new Mail_Queue_Error(MAILQUEUE_ERROR_QUERY_FAILED,
                 $this->pearErrorMode, E_USER_ERROR, __FILE__, __LINE__,
-                'MDB2::query failed - "'.$query.'" - '.MDB2::errorMessage($res));
+                sprintf($this->errorMsg, $query, $this->_getErrorMessage($res)));
         }
         return true;
     }
